# docs/swagger.yaml
definitions:
...
  model.CreateFacultyInput:
    properties:
      full_name:
        type: string
    type: object
  model.CreateGroupInput:
    properties:
      educational_level:
        type: string
      full_name:
        type: string
      specialty_id:
        type: string
      start_year:
        type: string
    type: object
...
  model.Faculty:
    properties:
      faculty_id:
        type: string
      full_name:
        type: string
    type: object
  model.Group:
    properties:
      educational_level:
        type: string
      end_year:
        type: string
      faculty_id:
        type: string
      full_name:
        type: string
      group_id:
        type: string
      specialty_id:
        type: integer
      start_year:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: REST API for test.
  title: University System API
  version: "2.0"
paths:
...
  /admin/hub/faculties:
    get:
      consumes:
      - application/json
      description: get faculties
      produces:
      - application/json
      responses:
        "200":
          description: Accepted
          schema:
            items:
              $ref: '#/definitions/model.Faculty'
            type: array
        "400":
          description: Bad request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get faculties
      tags:
      - admin-faculties
    post:
      consumes:
      - application/json
      description: create faculty
      parameters:
      - description: Data for creating faculty
        in: body
        name: faculty
        required: true
        schema:
          $ref: '#/definitions/model.CreateFacultyInput'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            type: string
        "400":
          description: Bad request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: create faculty
      tags:
      - admin-faculties
  /admin/hub/faculties/{id}:
    delete:
      consumes:
      - application/json
      description: delete faculty
      parameters:
      - description: ID for deleting faculty
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Delete faculty
      tags:
      - admin-faculties
    get:
      consumes:
      - application/json
      description: get faculty by id
      parameters:
      - description: ID for getting faculty
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Accepted
          schema:
            $ref: '#/definitions/model.Faculty'
        "400":
          description: Bad request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get faculty
      tags:
      - admin-faculties
  /admin/hub/groups:
    get:
      consumes:
      - application/json
      description: get groups
      produces:
      - application/json
      responses:
        "200":
          description: Accepted
          schema:
            items:
              $ref: '#/definitions/model.Group'
            type: array
        "400":
          description: Bad request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get groups
      tags:
      - admin-groups
    post:
      consumes:
      - application/json
      description: create group
      parameters:
      - description: Data for creating group
        in: body
        name: faculty
        required: true
        schema:
          $ref: '#/definitions/model.CreateGroupInput'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            type: string
        "400":
          description: Bad request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: create group
      tags:
      - admin-groups
  /admin/hub/groups/{id}:
    delete:
      consumes:
      - application/json
      description: delete group
      parameters:
      - description: ID for deleting group
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Delete group
      tags:
      - admin-groups
    get:
      consumes:
      - application/json
      description: get group by id
      parameters:
      - description: ID for getting group
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Accepted
          schema:
            $ref: '#/definitions/model.Faculty'
        "400":
          description: Bad request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get group
      tags:
      - admin-groups
...
swagger: "2.0"